requirements.txt
# Core dependencies
streamlit==1.31.1
python-dotenv==1.0.1

# LLM and embeddings
groq==0.4.2
sentence-transformers==2.3.1
langchain==0.1.10
langchain-community==0.0.25

# Document processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Vector database
chromadb==0.4.22
sentence-transformers==2.3.1

# Utilities
numpy==1.26.4
pandas==2.2.0
tqdm==4.66.1
tiktoken==0.5.2

# Evaluation
scikit-learn==1.4.1.post1
